refactor: 패키지 이동 및 매직넘버 제거
