--- tomcat/src/main/java/org/apache/coyote/http11/http/HttpMethod.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.http11.http;
-
-public enum HttpMethod {
-
-    GET,
-}
