+++ tomcat/src/main/java/org/apache/coyote/http11/http/HttpRequest.java
@@ -2,6 +2,9 @@ package org.apache.coyote.http11.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import org.apache.coyote.http11.http.domain.Headers;
+import org.apache.coyote.http11.http.domain.MessageBody;
+import org.apache.coyote.http11.http.domain.RequestLine;
 
 public class HttpRequest {
