+++ tomcat/src/main/java/org/apache/coyote/http11/http/HttpResponse.java
@@ -1,5 +1,13 @@
 package org.apache.coyote.http11.http;
 
+import org.apache.coyote.http11.http.domain.ContentType;
+import org.apache.coyote.http11.http.domain.Headers;
+import org.apache.coyote.http11.http.domain.HttpConstants;
+import org.apache.coyote.http11.http.domain.HttpVersion;
+import org.apache.coyote.http11.http.domain.MessageBody;
+import org.apache.coyote.http11.http.domain.StatusCode;
+import org.apache.coyote.http11.http.domain.StatusLine;
+
 public class HttpResponse {
 
     private final StatusLine statusLine;
