+++ tomcat/src/main/java/org/apache/coyote/http11/http/domain/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.http;
+package org.apache.coyote.http11.http.domain;
 
 import java.util.Arrays;
 import nextstep.jwp.exception.NotFoundException;
