+++ tomcat/src/main/java/org/apache/coyote/http11/http/domain/MessageBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.http;
+package org.apache.coyote.http11.http.domain;
 
 import java.io.BufferedReader;
 import java.io.IOException;
