+++ tomcat/src/main/java/org/apache/coyote/http11/http/domain/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.http;
+package org.apache.coyote.http11.http.domain;
 
 public enum StatusCode {
