+++ tomcat/src/test/java/org/apache/coyote/http11/http/ContentTypeTest.java
@@ -3,6 +3,7 @@ package org.apache.coyote.http11.http;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import org.apache.coyote.http11.http.domain.ContentType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
