+++ tomcat/src/test/java/org/apache/coyote/http11/http/MessageBodyTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import org.apache.coyote.http11.http.domain.MessageBody;
 import org.junit.jupiter.api.Test;
 import support.BufferedReaderFactory;
