+++ tomcat/src/test/java/org/apache/coyote/http11/http/RequestLineTest.java
@@ -5,6 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import org.apache.coyote.http11.http.domain.HttpMethod;
+import org.apache.coyote.http11.http.domain.HttpVersion;
+import org.apache.coyote.http11.http.domain.RequestLine;
 import org.junit.jupiter.api.Test;
 import support.BufferedReaderFactory;
