+++ tomcat/src/test/java/org/apache/coyote/http11/http/StatusCodeTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.http.domain.StatusCode;
 import org.junit.jupiter.api.Test;
 
 class StatusCodeTest {
