+++ tomcat/src/test/java/org/apache/coyote/http11/http/StatusLineTest.java
@@ -2,6 +2,9 @@ package org.apache.coyote.http11.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.http.domain.HttpVersion;
+import org.apache.coyote.http11.http.domain.StatusCode;
+import org.apache.coyote.http11.http.domain.StatusLine;
 import org.junit.jupiter.api.Test;
 
 class StatusLineTest {
